+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - WireGuard support. [#236](https://github.com/passepartoutvpn/tunnelkit/pull/236)
+- Handle `--keepalive` option.
 
 ## 4.0.3 (2021-11-27)